+++ src/markadoros/bold_tsv_processor.py
@@ -77,7 +77,7 @@ class BoldTSVProcessor:
         )
 
         logger.info(f"Processing BOLD TSV file with Polars: {tsv}")
-        df = (
+        df_lazy = (
             pl.scan_csv(
                 tsv,
                 separator="\t",
@@ -92,12 +92,38 @@ class BoldTSVProcessor:
                 & (pl.col("marker_code") == "COI-5P")
             )
             .group_by("bin_uri")
-            .map_groups(
+        )
+
+        df = (
+            df_lazy.map_groups(
                 lambda group: self.select_best_row(group, self.taxonomy_levels),
                 schema=schema,
             )
             .collect(engine="streaming")
         )
+
+        bins = (
+            df_lazy.agg(
+                pl.concat_list([
+                    pl.col("species"),
+                    pl.col("genus"),
+                    pl.col("tribe"),
+                    pl.col("subfamily"),
+                    pl.col("family"),
+                    pl.col("order"),
+                    pl.col("class"),
+                    pl.col("phylum"),
+                    pl.col("kingdom"),
+                ])
+                .list.drop_nulls()
+                .list.first()
+                .alias("taxonomy")
+            )
+            .collect(engine="streaming")
+        ).to_dicts()
+
+        bins_taxa = {row["bin_uri"]: row["taxonomy"] for row in bins}
+
         ## Make the type-checker happy
         assert isinstance(df, pl.DataFrame)
 
@@ -130,4 +156,4 @@ class BoldTSVProcessor:
             if collected:
                 fout.writelines(collected)
 
-        return outfile
+        return outfile, bins_taxa